Reprocess data with upstream fixes for 0 vs. “no value”, and site detection
For dates before 2014-09-23, there is no value for mobile and zero
sites. We used to mark that with 0, but it should be empty to help
plotting libraries distinguish between “no value” and 0.

Also the upstream fixes weed out outliers on the mobile site, and
allow to detect mobile site for foundationwiki and mediawikiwiki.

Bug: T76406
Change-Id: I5b3368ebf0739c5a60f09a5b0408b0fb494de97f